requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
et-xmlfile==1.0.1
idna==2.9
jdcal==1.4.1
openpyxl==3.0.3
requests==2.23.0
trello.py==0.0.6.post4
urllib3==1.25.9
